전체 정적 코드 검토 및 버그 수정

+++ lib/widgets/review_widget_factory.dart
@@ -31,7 +31,7 @@ class ReviewWidgetFactory extends WidgetFactory {
                 ...List.generate(
                   5,
                   (index) => Icon(
-                    index < review.score ? Icons.star : Icons.star_border,
+                    index < review.score.toInt() ? Icons.star : Icons.star_border,
                     color: context.starColor,
                     size: 20,
                   ),
